// src/generated/zod/userProfileResponse.zod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * One Health Lyme Gap Atlas API
 * Build artifact for existing first-party Web client and validator generation. Not an external developer contract or HTTP surface. Use the public openapi.json for external documentation.
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';

export const userProfileResponseProfileOneStateCodeOneMax = 2;

export const userProfileResponseProfileOneOrganizationOneMax = 120;

export const userProfileResponseProfileOneJobTitleOneMax = 120;


export const UserProfileResponse = zod.object({
  "profile": zod.union([zod.object({
  "role": zod.union([zod.enum(['general_public_citizen', 'district_level_epidemiologist', 'state_level_epidemiologist', 'state_director_level_epidemiologist', 'national_level_epidemiologist']),zod.null()]).optional(),
  "state_code": zod.union([zod.string().max(userProfileResponseProfileOneStateCodeOneMax),zod.null()]).optional(),
  "organization": zod.union([zod.string().max(userProfileResponseProfileOneOrganizationOneMax),zod.null()]).optional(),
  "job_title": zod.union([zod.string().max(userProfileResponseProfileOneJobTitleOneMax),zod.null()]).optional()
}),zod.null()])
})

export type UserProfileResponse = zod.input<typeof UserProfileResponse>;
export type UserProfileResponseOutput = zod.output<typeof UserProfileResponse>;