// src/generated/zod/userProfileWrite.zod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * One Health Lyme Gap Atlas API
 * Build artifact for existing first-party Web client and validator generation. Not an external developer contract or HTTP surface. Use the public openapi.json for external documentation.
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';

export const userProfileWriteStateCodeOneMax = 2;

export const userProfileWriteOrganizationOneMax = 120;

export const userProfileWriteJobTitleOneMax = 120;


export const UserProfileWrite = zod.object({
  "role": zod.union([zod.enum(['general_public_citizen', 'district_level_epidemiologist', 'state_level_epidemiologist', 'state_director_level_epidemiologist', 'national_level_epidemiologist']),zod.null()]).optional(),
  "state_code": zod.union([zod.string().max(userProfileWriteStateCodeOneMax),zod.null()]).optional(),
  "organization": zod.union([zod.string().max(userProfileWriteOrganizationOneMax),zod.null()]).optional(),
  "job_title": zod.union([zod.string().max(userProfileWriteJobTitleOneMax),zod.null()]).optional()
})

export type UserProfileWrite = zod.input<typeof UserProfileWrite>;
export type UserProfileWriteOutput = zod.output<typeof UserProfileWrite>;